+++ src/elements/GeoChart/GeoChart.vue
@@ -281,6 +281,13 @@ export default {
       deep: true
     },
 
+    'config.quadrantGroups': {
+      handler () {
+        this.debouncedRedraw()
+      },
+      deep: true
+    },
+
     'config.lineGroups': {
       handler () {
         this.debouncedRedraw()
@@ -313,6 +320,7 @@ export default {
       this.adjustSize()
       this.redrawGuidelines()
       this.updateData()
+      this.redrawQuadrants()
       this.redrawAxes()
     },