+++ src/elements/GeoChart/GeoChartScatterPlot/GeoChartScatterPlot.d.ts
@@ -3,6 +3,7 @@ declare namespace GeoChart {
     data: [number, number][]
     radius?: number
     fillColor?: string
+    onDotClick?: (item: object, index: number) => string | null | undefined
     tooltip?: {
       getContent?: (item: object, index: number) => string | null | undefined
       getOffset?: (event: MouseEvent) => { x: number, y: number } | null | undefined